cronjobs/redditad/requirements.txt
certifi==2020.12.5
chardet==4.0.0
ConfigArgParse==1.4
idna==2.10
praw==7.4.0
prawcore==2.3.0
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.59.0
